src/requirements.txt
fastapi==0.115.11
pydantic_settings==2.8.1
sqlalchemy==2.0.39
alembic==1.15.1
uvicorn==0.34.0
asyncpg==0.30.0
black==25.1.0
python-multipart==0.0.20
cffi==1.17.1 
cryptography==44.0.2 
pycparser==2.22 
pyjwt==2.10.1
bcrypt==4.3.0
dnspython==2.7.0 
email-validator==2.2.0
redis==5.2.1 
slowapi==0.1.9
# botocore==1.37.20
boto3==1.37.20